Try to mop up failures from previous big reorg

+++ src/core/include/tile.tcc
@@ -325,13 +325,13 @@ void Tile<T,R>::IncrAtomCount(ElementType atomType, s32 delta)
     if (delta != 0) FAIL(ILLEGAL_STATE);  
     return;
   }
-  if (delta < 0 && (u32) -delta > m_atomCount[idx])
+  if (delta < 0 && -delta > m_atomCount[idx])
     FAIL(ILLEGAL_ARGUMENT);
 
   if (delta < 0)
-    m_atomCount[idx] -= (u32) -delta;
+    m_atomCount[idx] -= -delta;
   else
-    m_atomCount[idx] += (u32) delta;
+    m_atomCount[idx] += delta;
 
 }
 } /* namespace MFM */
